moving context and plugin objects from config to utils so config can be used in another place
