+++ lib/assemble.js
@@ -45,7 +45,6 @@ var Assemble = function (src, options) {
   this.appName = 'assemble';
 
   // add assemble "static" objects
-  this.config   = Assemble.config;
   this.defaults = Assemble.defaults;
   this.models   = Assemble.models;
   this.render   = Assemble.render;
@@ -75,7 +74,7 @@ var Assemble = function (src, options) {
   this.middleware = [];
 
   // setup the context
-  this.context = this.config.context.stub(this);
+  this.context = this.utils.context.stub(this);
 
 };
 require('util').inherits(Assemble, events.EventEmitter);
@@ -85,11 +84,10 @@ Assemble.render = require('./render');
 Assemble.logger = Assemble.utils.log;
 
 var middleware = require('./middleware');
-var config = Assemble.config = require('./config');
 Assemble.models = require('./models');
 
 // add assemble plugin events to Dovetail
-Dovetail.plugins.addEvents(config.plugins.events);
+Dovetail.plugins.addEvents(Assemble.utils.plugins.events);
 
 Assemble.paths = {
   defaults: relative(path.join(__dirname, 'defaults.yml')),
@@ -121,7 +119,7 @@ Assemble.prototype.use = function(middleware) {
 Assemble.prototype.build = function (callback) {
 
   // reset the context before building
-  this.context = this.config.context.stub(this);
+  this.context = this.utils.context.stub(this);
 
   // loop over each middleware
   async.eachSeries(
