--- lib/config/index.js
@@ -1,13 +1,0 @@
-/**
- * Assemble <http://assemble.io>
- *
- * Copyright (c) 2014 Jon Schlinkert, Brian Woodward, contributors
- * Licensed under the MIT License (MIT).
- */
-
-'use strict';
-
-var config = module.exports = {};
-
-config.plugins = require('./plugins');
-config.context = require('./context');
