+++ lib/middleware/build.js
@@ -14,12 +14,12 @@ var _ = require('lodash');
 
 // Local libs
 var notifier = require('./notifier');
-var config = require('../config');
+var utils = require('../utils');
 var registerPartials = require('./partials');
 var registerComponents = require('./components');
 var registerPages = require('./pages');
 
-var events = config.plugins.events;
+var events = utils.plugins.events;
 
 var normalizePages = function (assemble) {
   var results = [];
