+++ lib/middleware/config.js
@@ -13,11 +13,11 @@ var async = require('async');
 var _ = require('lodash');
 
 // Local libs
-var config = require('../config');
+var utils = require('../utils');
 var notifier = require('./notifier');
 var registerMixins = require('./mixins');
 var Component = require('../models/component');
-var events = config.plugins.events;
+var events = utils.plugins.events;
 
 module.exports = function(assemble, done) {
