+++ lib/middleware/data.js
@@ -14,9 +14,9 @@ var _ = require('lodash');
 
 // Local libs
 var notifier = require('./notifier');
-var config = require('../config');
+var utils = require('../utils');
 
-var events = config.plugins.events;
+var events = utils.plugins.events;
 
 module.exports = function (assemble, done) {
