+++ lib/middleware/pages.js
@@ -2,9 +2,9 @@
 var async = require('async');
 
 var utils = require('../utils');
-var config = require('../config');
+var utils = require('../utils');
 var notifier = require('./notifier');
-var events = config.plugins.events;
+var events = utils.plugins.events;
 
 /**
  * Register pages
