+++ lib/middleware/render.js
@@ -13,9 +13,9 @@ var _ = require('lodash');
 
 // Local libs
 var notifier = require('./notifier');
-var config = require('../config');
+var utils = require('../utils');
 
-var events = config.plugins.events;
+var events = utils.plugins.events;
 
 module.exports = function(assemble, done) {
 
@@ -47,7 +47,7 @@ module.exports = function(assemble, done) {
           // add the current page to the params
           // so plugins and make modifications if needed.
           params.page = assemble.pages[pageKey];
-          config.context(assemble, params);
+          utils.context(assemble, params);
 
           // run steps in series so we can notify plugins
           // before and after the page steps are done.
@@ -95,7 +95,7 @@ module.exports = function(assemble, done) {
       }
 
       params.page = assemble.source;
-      config.context(assemble, params);
+      utils.context(assemble, params);
 
       async.series([
         notify(events.pageBeforeRender),
