+++ lib/plugins/core-collections.js
@@ -19,7 +19,7 @@ var _ = require('lodash');
 
 module.exports = function(assemble) {
 
-  var events = assemble.config.plugins.events;
+  var events = assemble.utils.plugins.events;
 
   var configureCollections = function (params, done) {
     assemble.log.debug('\t[core plugin]: ', 'core-collections-config plugin', params.event);
