+++ lib/plugins/core-components.js
@@ -11,7 +11,7 @@
 
 module.exports = function(assemble) {
 
-  var events = assemble.config.plugins.events;
+  var events = assemble.utils.plugins.events;
 
   var plugin = function (params, done) {
     assemble.log.debug('\t[core plugin]: ', 'core-component building plugin', params.event);
