+++ lib/plugins/core-pagination-step1.js
@@ -13,7 +13,7 @@ var _ = require('lodash');
 
 module.exports = function(assemble) {
 
-  var events = assemble.config.plugins.events;
+  var events = assemble.utils.plugins.events;
 
   var plugin = function (params, done) {
     assemble.log.debug('\t[core plugin]: ', 'pagination plugin: step 1', params.event);
