+++ lib/plugins/core-pagination-step2.js
@@ -12,7 +12,7 @@ var _ = require('lodash');
 
 module.exports = function(assemble) {
 
-  var events = assemble.config.plugins.events;
+  var events = assemble.utils.plugins.events;
 
   var plugin = function (params, done) {
     assemble.log.debug('\t[core plugin]: ', 'pagination plugin: step 2', params.event);
