+++ lib/plugins/core-paths.js
@@ -16,7 +16,7 @@ var _ = require('lodash');
 
 module.exports = function(assemble) {
 
-  var events = assemble.config.plugins.events;
+  var events = assemble.utils.plugins.events;
 
   var plugin = function (params, done) {
     assemble.log.debug('\t[core plugin]: ', 'info plugin', params.event);
