+++ lib/utils/context.js
@@ -12,7 +12,7 @@
 var _ = require('lodash');
 
 // local modules
-var data = require('../utils').data;
+var data = require('./data');
 
 /**
  * Add context functions to be used on params
